requirements.txt
langchain==1.0.2
langchain-openai==1.0.1
langchain-mcp-adapters>=0.1.0
fastmcp==2.12.5
pyyaml

# A_stock
tushare
efinance

# Scheduled trading
apscheduler>=3.10.0
pytz
akshare

# Database
duckdb>=1.0.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
